Add remote signature source support

+++ src/main/java/org/iota/compass/crypto/ISS.java
@@ -41,7 +41,7 @@ public class ISS {
   private static final int MIN_TRIT_VALUE = -1, MAX_TRIT_VALUE = 1;
   private static final int MIN_TRYTE_VALUE = -13, MAX_TRYTE_VALUE = 13;
 
-  public static int[] subseed(SpongeFactory.Mode mode, final int[] seed, int index) {
+  public static int[] subseed(SpongeFactory.Mode mode, final int[] seed, long index) {
 
     if (index < 0) {
       throw new RuntimeException("Invalid subseed index: " + index);
